# swagger.yaml
openapi: 3.0.0
info:
  title: Medbox API
  description: MedBox API
  version: 1.0.0
servers:
  - url: http://localhost:4000
    description: Local development server
paths:
...
  /accounts/refresh-token:
    post:
      summary: Use a refresh token to generate a new JWT token and a new refresh token
      description: The refresh token is sent and returned via cookies.
      operationId: refreshToken
      parameters:
        - in: cookie
          name: refreshToken
          description: The `refreshToken` cookie
          schema:
            type: string
            example: 51872eca5efedcf424db4cf5afd16a9d00ad25b743a034c9c221afc85d18dcd5e4ad6e3f08607550
      responses:
        "200":
          description: Account details, a JWT access token and a new refresh token cookie
          headers:
            Set-Cookie:
              description: "`refreshToken`"
              schema:
                type: string
                example: refreshToken=51872eca5efedcf424db4cf5afd16a9d00ad25b743a034c9c221afc85d18dcd5e4ad6e3f08607550; Path=/; Expires=Tue, 16 Jun 2020 09:14:17 GMT; HttpOnly
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    example: "5eb12e197e06a76ccdefc121"
                  title:
                    type: string
                    example: "Mr"
                  firstName:
                    type: string
                    example: "Jason"
                  lastName:
                    type: string
                    example: "Watmore"
                  email:
                    type: string
                    example: "jason@example.com"
                  role:
                    type: string
                    example: "Admin"
                  created:
                    type: string
                    example: "2020-05-05T09:12:57.848Z"
                  isVerified:
                    type: boolean
                    example: true
                  jwtToken:
                    type: string
                    example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiI1ZWIxMmUxOTdlMDZhNzZjY2RlZmMxMjEiLCJpZCI6IjVlYjEyZTE5N2UwNmE3NmNjZGVmYzEyMSIsImlhdCI6MTU4ODc1ODE1N30.xR9H0STbFOpSkuGA9jHNZOJ6eS7umHHqKRhI807YT1Y"
        "400":
          description: The refresh token is invalid, revoked or expired
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Invalid token"
...
  /publicities:
    get:
      summary: Get a list of all publicities
      description: Restricted to admin users.
      operationId: getAllPublicities
      security:
        - bearerAuth: [ ]
      responses:
        "200":
          description: An array of all publicities
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: string
                      example: "5eb12e197e06a76ccdefc121"
                    description:
                      type: string
                      example: "Lorem ipsum Lorem ipsum Lorem ipsum Lorem ipsum Lorem ipsum Lorem ipsum Lorem ipsum Lorem ipsum Lorem ipsum"
                    image:
                      type: string
                      example: "path/to/image"
                    isActive:
                      type: boolean
                      example: fasle

        "401":
          $ref: "#/components/responses/UnauthorizedError"
    post:
      summary: Create a new publicities
      description: Restricted to admin users.
      operationId: createPublicity
      security:
        - bearerAuth: [ ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                description:
                  type: string
                  example: "Lorem ipsum Lorem ipsum Lorem ipsum Lorem ipsum Lorem ipsum Lorem ipsum Lorem ipsum Lorem ipsum Lorem ipsum"
                image:
                  type: string
                  example: "path/to/image"
              required:
                - description
                - image
      responses:
        "200":
          description: Publicity created successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    example: "5eb12e197e06a76ccdefc121"
                  description:
                    type: string
                    example: "Lorem ipsum Lorem ipsum Lorem ipsum Lorem ipsum Lorem ipsum Lorem ipsum Lorem ipsum"
                  image:
                    type: string
                    example: "path/to/image"
                  isActive:
                    type: boolean
                    example: "false"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
  /publicities/{id}:
    parameters:
      - in: path
        name: id
        description: Publicity id
        required: true
        example: "5eb12e197e06a76ccdefc121"
        schema:
          type: string
    get:
      summary: Get a single publicity by id
      description: Admin users can access any publicity
      operationId: getPublicityById
      security:
        - bearerAuth: [ ]
      responses:
        "200":
          description: Details of the specified publicity
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    example: "5eb12e197e06a76ccdefc121"
                  description:
                    type: string
                    example: "lorem ipsum lorem ipsum"
                  image:
                    type: string
                    example: "path/to/image"

        "404":
          $ref: "#/components/responses/NotFoundError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
    put:
      summary: Update a publicity
      description: Admin users can update any pub
      operationId: updatePublicity
      security:
        - bearerAuth: [ ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                description:
                  type: string
                  example: "Lorem lorem"
                image:
                  type: string
                  example: "path/to/image"
                isActive:
                  type: boolean
                  example: true

      responses:
        "200":
          description: publicity updated successfully. The details of the updated publicity are returned.
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    example: "5eb12e197e06a76ccdefc121"
                  description:
                    type: string
                    example: "lorem lorem"
                  image:
                    type: string
                    example: "path/to/image"

        "404":
          $ref: "#/components/responses/NotFoundError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
    delete:
      summary: Delete a publicity
      description: Admin users can delete any pub,
      operationId: deletePublicity
      security:
        - bearerAuth: [ ]
      responses:
        "200":
          description: Publicity deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "publicity deleted successfully"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
  /publicities/active:
    get:
      summary: Get a list of active publicities
      description: Restricted to admin users.
      operationId: getActivePublicities
      security:
        - bearerAuth: [ ]
      responses:
        "200":
          description: An array of active publicities
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: string
                      example: "5eb12e197e06a76ccdefc121"
                    description:
                      type: string
                      example: "Lorem ipsum Lorem ipsum Lorem ipsum Lorem ipsum Lorem ipsum Lorem ipsum Lorem ipsum Lorem ipsum Lorem ipsum"
                    image:
                      type: string
                      example: "path/to/image"
                    isActive:
                      type: boolean
                      example: true

        "401":
          $ref: "#/components/responses/UnauthorizedError"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    UnauthorizedError:
      description: Access token is missing or invalid, or the user does not have access to perform the action
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: "Unauthorized"
    NotFoundError:
      description: Not Found
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: "Not Found"